infrastructure/modules/functions/slack_notice/src/requirements.txt
slackclient
jpholiday